recognite_font/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.66.0
pillow>=9.0.0
scikit-learn>=1.7.0
numpy>=1.26.0